added themes support and various fixes

+++ daemon/wmanager/wmanager.h
@@ -19,6 +19,7 @@
 
 #include <stdint.h>
 #include <stdbool.h>
+#include "config.h"
 
 enum window_flags {
 	LISTEN_EVENTS = 0x1,
@@ -32,13 +33,13 @@ struct window_t {
 	int flags;
 	int x, y;
 	int width, height;
-	uint8_t *bitmap;
+	uint32_t *bitmap;
 	bool mutex;
 	struct window_t *next, *prev;
 };
 
 extern uint64_t vgafd;
-extern uint8_t *screen;
+extern uint32_t *screen;
 extern int screen_width, screen_height;
 extern struct window_t *windows;
 extern struct window_t *active_window;
@@ -57,7 +58,7 @@ void update_decorations(struct window_t *window);
 
 void resize_screen(int width, int height);
 void update_screen(int x1, int y1, int x2, int y2);
-void blit_bitmap(const uint8_t *bitmap, int tox, int toy, int width, int height, int x1, int y1, int x2, int y2);
+void blit_bitmap(const uint32_t *bitmap, int tox, int toy, int width, int height, int x1, int y1, int x2, int y2);
 
 void mouse_move(int16_t dx, int16_t dy);
 void mouse_buttons(int buttons);
